Conflicts:
	app/Providers/KeystoneGuruServiceProvider.php
	app/Service/View/ViewService.php
	app/Service/View/ViewServiceInterface.php
	tests/Feature/Mapping/DefaultFloorTest.php

+++ app/Service/Controller/Api/V1/APIDungeonRouteControllerService.php
@@ -14,7 +14,7 @@ class APIDungeonRouteControllerService implements APIDungeonRouteControllerServi
     }
 
     /**
-     * @return Collection<DungeonRouteThumbnailJob>
+     * @return Collection<int, DungeonRouteThumbnailJob>
      */
     public function createThumbnails(
         DungeonRoute $dungeonRoute,
@@ -22,7 +22,7 @@ class APIDungeonRouteControllerService implements APIDungeonRouteControllerServi
         ?int         $viewportHeight = null,
         ?int         $imageWidth = null,
         ?int         $imageHeight = null,
-        ?int         $zoomLevel = null,
+        ?float       $zoomLevel = null,
         ?int         $quality = null,
     ): Collection {
         return $this->thumbnailService->queueThumbnailRefreshForApi(
